Promote the old Gemini using the SDK as "gemini_legacy"
