+++ core/aimodels/gemini/helper_methods.py
